feat(pack): Reorganizing folder
