+++ module/pack/__index__.ts
@@ -1,2 +1,2 @@
-export type { CommonPackConfig, DockerPackFactory, DockerPackConfig } from './support/bin/types';
-export { PackConfigUtil } from './support/bin/config-util';
+export type { CommonPackConfig, DockerPackFactory, DockerPackConfig } from './src/types';
+export { PackConfigUtil } from './src/config-util';
