+++ module/pack/src/config-util.ts
